+++ CHANGELOG.md
@@ -12,6 +12,8 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
 - Basic hercules configuration support.
 
+- Templates no longer have bash prompt override.
+
 ## 2.0.1 -- 2022-12-12
 
 - Vendored version of mkHackage instead of using now deprecated `haskell-nix-extra-hackage`.